Changes in functions. Fixed names. localStorage.
Changes in main.js and all its modules. Removal of BuildArray in main.js to make it more dynamic, modification of functions next, previous, search, ids arrangement in Html and Css, modification of validationModule.js, modification of variable and function names, removal of await inside for, inclusion of Promise.all. Incorporation of localStorage.